Incorporate reagents with related entities, drop material column from lots table
Defines relationships with other tables and incorporates reagents into UI where applicable.

Updates QuickForm component to allow passed field headers to be a function that will be passed a copy of the current record for calculating custom field headers.  First use case implemented here shows the concentration units for the associated reagent when recording reagent usage in the UI.