Retain object as named model value in AutoCompleter component and add ref to store these as related records in QuickForm
In some cases, the values from the underlying object of AutoCompleter component are needed within the QuickForm to label other parts of the form. To accomplish this, the underlying objects are now a named model value, which become part of the QuickForm relatedRecords object. The relatedRecords object can then be used in other parts of the QuickForm; the initial use case is also included on the lots page.