Adding feature to export targets TSV file from genes page
New button added to each row of the genes table will export a targets TSV file with the format needed for external pipeline. All targets associated with the selected gene are included in the output.

Also added the relationships needed for the underlying drizzle query.

The endpoint is currently housed under 'custom' route path because adding a 'genes' folder will disrupt use of the generic endpoints for that table. It's possible this could be resolved by some custom Nuxt configuration, but for now will need to split generic and custom endpoints this way.