Fix DataTable column filtering in QuickTable component for calculated fields
Column-level filtering in DataTable component was not working for computed fields because the filtering and display were operating on diffferent values. This solves that issue by applying the formatting functions up front, before populating the data table, and storing that value as 'displayValue' property. This value is then treated as a string for display and filtering.