Handle checkboxes for boolean fields in QuickFormMultiple component
When dealing with multiple records, checkbox inputs require special handling for styling, labeling, and clearing values. Resulting behavior is similar to other types of inputs, but implementation is distinct, requring passthrough for styling, addition of a button to clear values, and a span element to show number of conflicting values rather than placeholder.