Add utility function for parsing PUT and POST request errors into more useful data objects
To handle various constrain errors and return errors with appropriate data objects, which fit the format expected on the UI side to display meaningful error messages, adding utility function for PUT and POST requests. This builds on the existing functionality by adding data objects for fields with unique constraints that contain the field name and message to display, in place of error message returned by the db.