Pass and return old values when updating PlateDiagram wells
To update related records in other components, passing and returning previous values of updated wells is required. For this use case in amplification primer plate page, the amplication primer rows in the included QuickTable component must be refreshed if they were referenced in either the old or new well contents.