QuickTable update: extand column definitions to allow sort feature to be disabled and add new element type
Updates to Column definitions utilized by QuickTable to allow sort feature to be disabled, and add a new type of 'element' which can pass set a custom html element as string or function returning a string. This is to allow for content other than text to be included in table cells, while keeping option open to allow any HTML content which is then imbedded via v-html attribute.