Define enum lookup const and apply during JSON Schema generation
Creates a enum lookup object to store short and long descriptions of enum values that are used in drizzle table and subsequenly zod object definitions. This can be used globally to define and retrieve descriptive values for any given enum coded value, and is keyed on table and field names.