Add support for view name as query param on GET requests
Allows for view names to be passed, to be queries directly in place of query builder queries. Views need to be pre-defined and should include all fields of the table/recordType/endpoint they are associated with plus any additional calculated fields. Because this bypasses the query builder, columns and with query params will be ignored, but the view definition may include its own conditional statements.