Replace plate diagram color map with well specs object
To optimize code, the plate diagram color map and related logic is being replace with a well specs object which stores all information about a well needed for the plate diagram, including color, symbol, and toolip. This structure will also allow for support for specs for wells with multiple contents, and centralize calculations in plate diagram utility function to avoid calcs in parent components.